requirements.txt
streamlit==1.44.0
SQLAlchemy==2.0.38
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
feedparser==6.0.11
google-genai==1.7.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
requests==2.32.3
